refactored wallet encryption to module
