+++ lib/WalletBackend.ts
@@ -4,16 +4,14 @@
 
 import { EventEmitter } from 'events';
 
-import * as crypto from 'crypto';
 import * as fs from 'fs';
 import * as _ from 'lodash';
-import * as pbkdf2 from 'pbkdf2';
 
 import { IDaemon } from './IDaemon';
 import { Metronome } from './Metronome';
 import { SubWallets } from './SubWallets';
 import { openWallet } from './OpenWallet';
-import { decryptWalletFromString } from './DecryptWallet';
+import { WalletEncryption } from './WalletEncryption';
 import { CryptoUtils} from './CnUtils';
 import { WalletBackendJSON } from './JsonSerialization';
 import { validateAddresses } from './ValidateParameters';
@@ -28,11 +26,7 @@ import {
     sendFusionTransactionAdvanced, sendFusionTransactionBasic,
 } from './Transfer';
 
-import {
-    IS_A_WALLET_IDENTIFIER, IS_CORRECT_PASSWORD_IDENTIFIER,
-    PBKDF2_ITERATIONS, WALLET_FILE_FORMAT_VERSION,
-    GLOBAL_INDEXES_OBSCURITY,
-} from './Constants';
+import { WALLET_FILE_FORMAT_VERSION, GLOBAL_INDEXES_OBSCURITY } from './Constants';
 
 import {
     addressToKeys, delay, getCurrentTimestampAdjusted, isHex64,
@@ -259,63 +253,13 @@ export class WalletBackend extends EventEmitter {
 
         MergeConfig(config);
 
-        const [walletJSON, error] = decryptWalletFromString(data, password);
+        const [walletJSON, error] = WalletEncryption.decryptWalletFromString(data, password);
 
         if (error) {
             return [undefined, error];
         }
 
         return WalletBackend.loadWalletFromJSON(deamon, walletJSON);
-    }
-
-    /**
-     * Encrypt the wallet using the given password. Note that an empty password does not mean an
-     * unencrypted wallet - simply a wallet encrypted with the empty string.
-     *
-     * This will take some time (Roughly a second on a modern PC) - it runs 500,000 iterations of pbkdf2.
-     *
-     * Example:
-     * ```javascript
-     * const dataJson = wallet.encryptWallet('hunter2');
-     *
-     * ```
-     *
-     * @param password The password to encrypt the wallet with
-     *
-     * @return Returns a JSON string containing the encrypted fileData.
-     */
-    public static encryptWallet(walletJson: string, password: string): Buffer {
-        /* Append the identifier so we can verify the password is correct */
-        const data: Buffer = Buffer.concat([
-            IS_CORRECT_PASSWORD_IDENTIFIER,
-            Buffer.from(walletJson),
-        ]);
-
-        /* Random salt */
-        const salt: Buffer = crypto.randomBytes(16);
-
-        /* PBKDF2 key for our encryption */
-        const key: Buffer = pbkdf2.pbkdf2Sync(password, salt, PBKDF2_ITERATIONS, 16, 'sha256');
-
-        /* Encrypt with AES */
-        const cipher = crypto.createCipheriv('aes-128-cbc', key, salt);
-
-        /* Perform the encryption */
-        const encryptedData: Buffer = Buffer.concat([
-            cipher.update(data),
-            cipher.final(),
-        ]);
-
-        /* Write the wallet identifier to the file so we know it's a wallet file.
-           Write the salt so it can be decrypted again */
-        const fileData: Buffer = Buffer.concat([
-            IS_A_WALLET_IDENTIFIER,
-            salt,
-            encryptedData,
-        ]);
-
-        return fileData;
-        // return JSON.stringify(fileData);
     }
 
     /**
@@ -1174,6 +1118,28 @@ export class WalletBackend extends EventEmitter {
     }
 
     /**
+     * encrypt the wallet using the given password. Password may be empty. Note that an empty password does not mean an
+     * unencrypted wallet - simply a wallet encrypted with the empty string.
+     *
+     * This will take some time (Roughly a second on a modern PC) - it runs 500,000 iterations of pbkdf2.
+     *
+     * Example:
+     * ```javascript
+     * const saved = wallet.encryptWalletToString('hunter2');
+     *
+     * ```
+     *
+     * @param password The password to encrypt the wallet with
+     *
+     * @return Returns the encrypted wallet as astring.
+     */
+    public encryptWalletToString(password: string) {
+        const walletJson: string = JSON.stringify(this);
+
+        return WalletEncryption.encryptWalletToString(walletJson, password);
+    }
+
+    /**
      * Save the wallet to the given filename. Password may be empty, but
      * filename must not be. Note that an empty password does not mean an
      * unencrypted wallet - simply a wallet encrypted with the empty string.
@@ -1196,7 +1162,7 @@ export class WalletBackend extends EventEmitter {
      */
     public saveWalletToFile(filename: string, password: string): boolean {
         const walletJson: string = JSON.stringify(this);
-        const fileData = WalletBackend.encryptWallet(walletJson, password);
+        const fileData = WalletEncryption.encryptWalletToBuffer(walletJson, password);
 
         try {
             fs.writeFileSync(filename, fileData);
